streamlit_app/pyproject.toml
[project]
name = "sony-pipeline-monitor"
version = "0.1.0"
description = "Sony Pictures Data Engineering Pipeline Monitor"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "snowflake-connector-python>=3.6.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
]
